Move model-utils main sources
