+++ tools/src/main/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/mapper/YugipediaCardTextMapper.java
@@ -13,8 +13,8 @@ import org.mapstruct.factory.Mappers;
 import io.github.ygojson.model.data.Card;
 import io.github.ygojson.model.data.definition.CardText;
 import io.github.ygojson.model.data.definition.localization.Language;
-import io.github.ygojson.model.utils.data.CardUtils;
-import io.github.ygojson.model.utils.data.utils.LocalizationUtils;
+import io.github.ygojson.model.utils.CardUtils;
+import io.github.ygojson.model.utils.LocalizationUtils;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.YugipediaException;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext.MarkupStringMapper;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.wikitext.CardTable2;
