+++ tools/src/main/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/mapper/YugipediaSetInfoMapper.java
@@ -14,7 +14,7 @@ import org.mapstruct.factory.Mappers;
 import io.github.ygojson.model.data.Set;
 import io.github.ygojson.model.data.definition.SetInfo;
 import io.github.ygojson.model.data.definition.localization.Language;
-import io.github.ygojson.model.utils.data.utils.LocalizationUtils;
+import io.github.ygojson.model.utils.LocalizationUtils;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.YugipediaException;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext.MarkupStringMapper;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.model.CardNumber;
